% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{plot.no.pattern}
\alias{plot.no.pattern}
\title{Generates plots for demo of package functions which take time series and
window width parameters}
\usage{
\method{plot}{no.pattern}(func, plt.title.vec)
}
\arguments{
\item{func}{runstats package core function}

\item{plt.title.vec}{vector of function-specific plot titles}
}
\value{
\code{NULL}
}
\description{
Generates plots for demo of package functions which take time series and
window width parameters
}
\examples{
\dontrun{
func <- RunningMean
vec <- c("black: x\\nred: W-width running window",
         "RunningMean(x, W)",
         "RunningMean(x, W, circular = TRUE)")
plot.no.pattern(func, vec)
}


}
