% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{plot.with.pattern}
\alias{plot.with.pattern}
\title{Generates plots for demo of package functions which take time series and
a short-timed pattern}
\usage{
\method{plot}{with.pattern}(func, plt.title.vec)
}
\arguments{
\item{func}{runstats package core function}

\item{plt.title.vec}{vector of function-specific plot titles}
}
\value{
\code{NULL}
}
\description{
Generates plots for demo of package functions which take time series and
a short-timed pattern
}
\examples{
\dontrun{
func <- RunningCov
vec <- c("black: x\\nred: running pattern",
         "RunningCov(x, pattern)",
         "RunningCov(x, pattern, circular = TRUE)")
plot.with.pattern(func, vec)
}


}
