% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{RunningMean}
\alias{RunningMean}
\title{Fast Running Mean Computation}
\usage{
RunningMean(x, W, circular = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{W}{A numeric scalar; width of \code{x} window over which sample mean is computed.}

\item{circular}{Logical; whether running sample mean is computed assuming
circular nature of  \code{x} sequence (see Details).}
}
\value{
A numeric vector.
}
\description{
Computes running sample mean of a sequence in a fixed width window. Uses
convolution implementation via Fast Fourier Transform.
}
\details{
The length of output vector equals the length of \code{x} vector.
Parameter \code{circular} determines whether \code{x} sequence is assumed to have a  circular nature.
Assume \eqn{l_x} is the length of sequence \code{x}, \code{W} is a fixed length of \code{x} sequence window.

If \code{circular} equals \code{TRUE} then
  \itemize{
    \item first element of the output sequence corresponds to sample mean of \code{x[1:W]},
    \item last element of the output sequence corresponds to sample mean of \code{c(x[l_x], x[1:(W - 1)])}.
  }

If \code{circular} equals \code{FALSE} then
  \itemize{
    \item first element of the output sequence corresponds to sample mean of \code{x[1:W]},
    \item \eqn{l_x - W + 1}-th element of the output sequence corresponds to
    sample mean of \code{x[(l_x - W + 1):l_x]},
    \item last \code{W-1} elements of the output sequence are filled with \code{NA}.
  }
}
\examples{
x <- rnorm(10)
RunningMean(x, 3, circular = FALSE)
RunningMean(x, 3, circular = TRUE)

}
