% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{runstats.demo}
\alias{runstats.demo}
\title{Demo visualization of package functions}
\usage{
runstats.demo(func.name = "Cov")
}
\arguments{
\item{func.name}{Character value; one of the following:
\itemize{
  \item "mean",
  \item "sd",
  \item "var",
  \item "cov",
  \item "cor",
  \item "l2norm".
}}
}
\value{
\code{NULL}
}
\description{
Generates demo visualization of output of methods for computing running
statistics.
}
\examples{
\dontrun{
runstats.demo(func.name = "cor")
runstats.demo(func.name = "cov")
runstats.demo(func.name = "l2norm")
runstats.demo(func.name = "mean")
runstats.demo(func.name = "sd")
runstats.demo(func.name = "var")
}

}
