% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{runstats.demo}
\alias{runstats.demo}
\title{Demo visualization of package functions}
\usage{
runstats.demo(func.name = "RunningCov")
}
\arguments{
\item{func.name}{Character value; one of the following:
\itemize{
  \item "RunningMean",
  \item "RunningSd",
  \item "RunningVar",
  \item "RunningCov",
  \item "RunningCor",
  \item "RunningL2Norm".
}}
}
\value{
\code{NULL}
}
\description{
Generates demo visualization of output of methods for computing running
statistics.
}
\examples{
\dontrun{
runstats.demo(func.name = "RunningMean")
runstats.demo(func.name = "RunningSd")
runstats.demo(func.name = "RunningVar")
runstats.demo(func.name = "RunningCov")
runstats.demo(func.name = "RunningCor")
runstats.demo(func.name = "RunningL2Norm")
}

}
